+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [2020-06-13] Allow for the two log levels to be configurable by [@lflare].
+- [2020-06-13] Added X-Cache header to image responses by [@lflare].
 
 ### Changed
 - [2020-06-13] Modified AsyncAppender queue size to 1024 by [@lflare].